+++ game/install.js
@@ -1,18 +1,17 @@
 'use strict';
 
-var names = Object.keys(servers), search, apps, _server, j;
+// Configure mailboxses
+var srv   = servers['65.32.48.22']; srv.chdir('/webroot/accounts');
+var boxes = srv.glob('*.account', ['only_files']), name, model = srv.readJSON('model.box'), account, write;
 
-for(var i = 0; i < names.length; i++) {
-  _server = servers[names[i]];
-  search  = _server.glob('apps/*', ['only_files']);
+for(var i = 0; i < boxes.length; i++) {
+  name    = boxes[i].vars[0];
+  write   = clone(model);
 
-  for(j = 0; j < search.length; j++) {
-    if(!servers.__store.dirExists('webroot/' + search[j].vars[0]))
-      fatal('Application "' + search[j].vars[0] + '" was not found on the store\'s server');
+  write.account       = srv.readJSON(boxes[i].path);
+  write.account.token = (server.generateId() + server.generateId()).replace(/\-/g, '');
 
-    if(!_server.importFolder(servers.__store.exportFolder('webroot/' + search[j].vars[0]), 'apps/' + search[j].vars[0]))
-      fatal('Failed to import application "' + search[j].vars[0] + '" to the server');
-
-    _server.removeFile(search[j].path);
-  }
+  srv.writeFile(boxes[i].path, write);
 }
+
+srv.removeFile('model.box');